# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ingest/runs": {
      "get": {
        "description": "Recent ingest runs (newest first). Caps `limit` to 200. Pass\n`job_name=...` to filter to a single job type \u2014 the /schedule\ndaily-MTD card uses this to fetch only its own runs.",
        "operationId": "list_ingest_runs_api_ingest_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "job_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ingest Runs",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/ingest/runs/{run_id}": {
      "get": {
        "description": "Single ingest_run row by id.",
        "operationId": "get_ingest_run_api_ingest_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingest Run",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/ingest/runs/{run_id}/templates/{template_key}/membership": {
      "get": {
        "description": "Universe membership captured by this run's templates phase, for\nthe given `template_key`. Reads `universe_id` + `this_month` from\nthe run's `templates_summary` array entry (set by the templates\nphase as each template's refresh completes). Returns 404 when the\nrun didn't include the requested template, or when its diff entry\nfailed (no `this_month`).",
        "operationId": "get_template_membership_for_run_api_ingest_runs__run_id__templates__template_key__membership_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "template_key",
            "required": true,
            "schema": {
              "title": "Template Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Template Membership For Run",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/ingest/scheduled-refresh/cron": {
      "post": {
        "description": "Cron entry point. Verifies `X-Cron-Secret`, inserts an `ingest_run`\nrow tagged `triggered_by='auto'`, spawns the work in a daemon thread,\nand returns the run_id immediately so the cron's curl exits fast.\nThe Railway cron just needs the 200 \u2014 it doesn't wait for completion.",
        "operationId": "cron_scheduled_refresh_api_ingest_scheduled_refresh_cron_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": false,
            "schema": {
              "default": "smart_daily",
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-cron-secret",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cron Scheduled Refresh",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/ingest/scheduled-refresh/trigger": {
      "post": {
        "description": "Manual trigger from the /schedule UI's Run-now button. Same work\nas the cron path, just `triggered_by='manual'`. No cron-secret \u2014\nauth is enforced by the frontend proxy middleware in `frontend/proxy.ts`.",
        "operationId": "trigger_scheduled_refresh_api_ingest_scheduled_refresh_trigger_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": false,
            "schema": {
              "default": "manual",
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Scheduled Refresh",
        "tags": [
          "ingest"
        ]
      }
    },
...
  }
}